Move getCollectionForm, getFormKey, makeInitializedForm to lib
