+++ static/js/flow/collectionTypes.js
@@ -16,3 +16,20 @@ export type Collection = CollectionListItem & {
 };
 
 export type CollectionList = Array<CollectionListItem>;
+
+export type CollectionFormState = {
+  key: ?string,
+  title: ?string,
+  description: ?string,
+  viewChoice: string,
+  viewLists: ?string,
+  adminChoice: string,
+  adminLists: ?string,
+};
+
+export type CollectionUiState = {
+  newCollectionForm: CollectionFormState,
+  editCollectionForm: CollectionFormState,
+  isNew: boolean,
+  selectedVideoKey: ?string
+};
