+++ static/js/lib/collection_test.js
@@ -1,8 +1,18 @@
 // @flow
 import { assert } from 'chai';
 
-import { getActiveCollectionDetail } from './collection';
+import {
+  getActiveCollectionDetail,
+  getCollectionForm,
+  getFormKey,
+  makeInitializedForm,
+} from './collection';
 import { makeCollection } from '../factories/collection';
+import { INITIAL_UI_STATE } from "../reducers/collectionUi";
+import {
+  PERM_CHOICE_NONE,
+  PERM_CHOICE_LISTS,
+} from "./dialog";
 
 import type { Collection } from "../flow/collectionTypes";
 import type { RestState } from "../flow/restTypes";
@@ -37,6 +47,46 @@ describe('collection library function', () => {
       it(`returns null when the state has ${testDescriptor}`, () => {
         assert.isNull(getActiveCollectionDetail(state));
       });
+    });
+  });
+
+  for (const isNew of [true, false]) {
+    describe(`with isNew = ${String(isNew)}`, () => {
+      it('getFormKey returns the key for the form', () => {
+        assert.equal(getFormKey(isNew), isNew ? "newCollectionForm" : "editCollectionForm");
+      });
+
+      it('getCollectionForm gets the expected form', () => {
+        const collectionUi = INITIAL_UI_STATE;
+        const key = isNew ? "newCollectionForm" : "editCollectionForm";
+        // this is explicitly comparing identity, not value equality
+        assert.isTrue(getCollectionForm(collectionUi) === collectionUi[key]);
+      });
+    });
+  }
+
+  it('makes a new form without a collection', () => {
+    assert.deepEqual(makeInitializedForm(), {
+      key: '',
+      title: '',
+      description: '',
+      adminChoice: PERM_CHOICE_NONE,
+      adminLists: "",
+      viewChoice: PERM_CHOICE_NONE,
+      viewLists: "",
+    });
+  });
+
+  it('makes a new form with an existing collection', () => {
+    let collection = makeCollection();
+    assert.deepEqual(makeInitializedForm(collection), {
+      key: collection.key,
+      title: collection.title,
+      description: collection.description,
+      adminChoice: PERM_CHOICE_LISTS,
+      adminLists: collection.admin_lists.join(","),
+      viewChoice: PERM_CHOICE_LISTS,
+      viewLists: collection.view_lists.join(","),
     });
   });
 });
